+++ src/main/java/com/ogc/standard/bo/impl/ExchangeCurrencyBOImpl.java
@@ -1,5 +1,6 @@
 package com.ogc.standard.bo.impl;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -132,7 +133,7 @@ public class ExchangeCurrencyBOImpl extends PaginableBOImpl<ExchangeCurrency>
     }
 
     @Override
-    public String applyExchange(User user, Long fromAmount,
+    public String applyExchange(User user, BigDecimal fromAmount,
             String fromCurrency, String toCurrency) {
         String code = OrderNoGenerater
             .generate(EGeneratePrefix.EXCHANGE_CURRENCY.getCode());
